Move all RC int values to long, to mirror RC backend (#6219 => v2) (#6228)
## Summary of changes

RC backend use long for there integer fields.

Backport of 
- #6219

+++ tracer/src/Datadog.Trace/RemoteConfigurationManagement/Protocol/Tuf/TargetCustom.cs
@@ -11,6 +11,6 @@ namespace Datadog.Trace.RemoteConfigurationManagement.Protocol.Tuf
     internal class TargetCustom
     {
         [JsonProperty("v")]
-        public int V { get; set; }
+        public long V { get; set; }
     }
 }
